Discord_Webhook_Keylogger_Payload
This script will keylog all keystrokes made by the logged in user on the target machine. It will send the results over once the keyboard is inactive for 10 seconds and will not send anything if the keyboard is not active. Simply replace the WEBHOOK_HERE in the script with the URL of your desired webook inside the '.